// src/resources/Add_productsUserInterface.ts
export enum Add_productsUserInterface {

    input_search = '//*[@id="search"]/input',
    button_search = '//*[@id="search"]/span/button',
    button_product = '//*[@id="content"]/div[3]/div/div/div[1]/a/img',
    input_quantity = '//*[@id="input-quantity"]',
    button_add = '//*[@id="button-cart"]',
    button_shoppingcart = '//*[@id="top-links"]/ul/li[4]/a/span',
    button_checkout = '//*[@id="content"]/div[3]/div[2]/a',
    button_billingdetails = '//*[@id="button-payment-address"]',
    button_deliverydetails = '//*[@id="button-shipping-address"]',
    button_deliverymethod = '//*[@id="button-shipping-method"]',
    check_terms_conditions = '//*[@id="collapse-payment-method"]/div/div[3]/div/input[1]',
    button_paymentmethod = '//*[@id="button-payment-method"]',
    button_confirm_order = '//*[@id="button-confirm"]',
    continue_home = '//*[@id="content"]/div/div/a'
}